Moved functions to different modules
